+++ base/src/api/actions/list.rs
@@ -40,7 +40,7 @@ pub fn queue(request: HttpRequest) -> Result<impl Responder> {
         .store
         .with_transaction(|tx| {
             let mut actions = Vec::new();
-            let iter = tx.actions().finished(span.context().clone())?;
+            let iter = tx.actions().queue(span.context().clone())?;
             for action in iter {
                 actions.push(action?);
             }