+++ base/src/store/backend/sqlite3/action.rs
@@ -48,7 +48,8 @@ SELECT
     state,
     state_payload
 FROM actions_history
-WHERE action_id = ?;
+WHERE action_id = ?
+ORDER BY time DESC;
 "#;
 const ACTION_INSERT: &str = "action.insert";
 const ACTION_INSERT_SQL: &str = r#"