+++ mongodb/src/version/v3_2/replica.rs
@@ -42,6 +42,10 @@ impl Agent for ReplicaSet {
         ))
     }
 
+    fn service_name(&self) -> String {
+        "mongod".into()
+    }
+
     fn shards(&self, span: &mut Span) -> Result<Shards> {
         self.common.shards(span)
     }